validate the id

+++ src/api/services/todo.service.js
@@ -14,7 +14,11 @@ const createTodo = async (todoBody) => {
 };
 
 const getTodoById = async (id) => {
-  return Todo.findById(id);
+  const todo = Todo.findById(id);
+  if (!todo) {
+    throw new ApiError(httpStatus.NOT_FOUND, "Todo not found");
+  }
+  return todo;
 };
 
 const getAllTodos = async () => {
@@ -26,21 +30,28 @@ const filterTodosByStatus = async (status) => {
 };
 
 const updateTodo = async (id, todoBody) => {
+  const todo = Todo.findById(id);
+  if (!todo) {
+    throw new ApiError(httpStatus.NOT_FOUND, "Todo not found");
+  }
   return Todo.findByIdAndUpdate(id, todoBody, { new: true });
 };
 
 const updateTodoStatus = async (id, todostatus) => {
-  const todo = {
-    status: todostatus,
+  const todo = Todo.findById(id);
+  if (!todo) {
+    throw new ApiError(httpStatus.NOT_FOUND, "Todo not found");
   }
   return Todo.findByIdAndUpdate(id, todostatus, { new: true });
-  // const todoData = await Todo.create(todo);
-  // return todoData;
 };
 
 const deleteTodo = async (id) => {
-    return Todo.findByIdAndDelete(id);
-  };
+  const todo = Todo.findById(id);
+  if (!todo) {
+    throw new ApiError(httpStatus.NOT_FOUND, "Todo not found");
+  }
+  return Todo.findByIdAndDelete(id);
+};
 
 module.exports = {
   createTodo,
